add complete test that shift calculations work for acaggaca
specific notes: 

a  number of corrections needed to be made, specifically

camshift includes HG1 atoms in its structures and adds them if they are
missing. The side-chain chemical shift calculations expect HG1 atoms to
be present an give erroneous results if they aren't. Disulphide shift
corrections are then added for cytines with SG1's that are closer than
9A.

in xcamshift the disulphide shift corrections is applied when a
disulphide patch is present and there is an -SG-SG- bond. This means
that the chemical shifts for the side chain give incorrect results as
the shifts due to the HG---HB are not present and so the disulphide
shift correction has to be corrected to allow for this.

It should also be noted that the lack of a HG1 atom also affects the non
boned term so (small) corrections have to be made between the shifts
calculated by xcamshift and those in the camshift test data set.

Finally hydrogen bonds are not currently implemented so these have to be
corrected for when using the test data. These corrections will be
removed when the hydrogen bond potential is complete.


this closes #11 and #1