component list can return a modified  component list for translation
This is in preparation for translating non bonded coefficient lists
These use atom_type_ids to lookup data. The atom_type_ids are not always
contiguous (not all chem types are present in all structures). So we
will pack the gaps in the  array with dummy entries to avoid a costly
call to get_components_for_atom_id which has to do a search rather than
a direct lookup