the energy term cache in the energy calculator is now held by a ref
references to data converted to c pointers have to be strongly
held for at least the lifetime of the c pointer they are converted to.
While generally objects in xcamshift hold these references there are
cases especially in testing where this predicate can be violated. This
should be extended to all data structures where a c pointer is extracted
(see a future ticket)